Added sphere intersection

+++ Ray.cpp
@@ -2,7 +2,7 @@
 
 Ray::Ray(const Point3 &origin, const Vec3 &direction) : origin(origin), direction(direction) {}
 
-Point3 Ray::point_on_ray(float distance) {
+Point3 Ray::point_on_ray(double distance) {
 
     //Equation of a ray
     return  direction * distance + origin;
